fix: dedupe metrics pipelines and chart a11y config, drop dead code
